requirements.txt
picamera2
pycryptodome
